Utilização de inner class na validação do usuário.
